refactor: ChatRoom, ChatRoomUser, Message 패키지 분리
