--- src/main/java/com/teamkrews/chat/model/ChatRoom.java
@@ -1,24 +1,0 @@
-package com.teamkrews.chat.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-
-@Entity
-@Table(name = "chat_room")
-@Getter
-@Setter
-public class ChatRoom {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "chat_room_id")
-    private Long chatRoomId;
-
-    private Long userCnt;
-}
