--- src/main/java/com/teamkrews/chat/model/ChatRoomCreationDto.java
@@ -1,14 +1,0 @@
-package com.teamkrews.chat.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class ChatRoomCreationDto {
-        private String workspaceUUID;
-        private Long creatorUserId;
-        private List<Long> userIds;
-}
