--- src/main/java/com/teamkrews/chat/model/ChatRoomNewStateFalseDto.java
@@ -1,15 +1,0 @@
-package com.teamkrews.chat.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRoomNewStateFalseDto {
-    public  Long chatRoomUserId;
-    public  final Boolean newState = Boolean.FALSE;
-
-    public ChatRoomNewStateFalseDto(Long chatRoomUserId) {
-        this.chatRoomUserId = chatRoomUserId;
-    }
-}
