--- src/main/java/com/teamkrews/chat/model/ChatRoomNewStateTrueDto.java
@@ -1,15 +1,0 @@
-package com.teamkrews.chat.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRoomNewStateTrueDto {
-    public  Long chatRoomId;
-    public  final Boolean newState = Boolean.TRUE;
-
-    public ChatRoomNewStateTrueDto(Long chatRoomId) {
-        this.chatRoomId = chatRoomId;
-    }
-}
