--- src/main/java/com/teamkrews/chat/model/ChatRoomUser.java
@@ -1,43 +1,0 @@
-package com.teamkrews.chat.model;
-
-import com.teamkrews.User.model.User;
-import com.teamkrews.workspace.model.Workspace;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "chat_room_user")
-@Getter
-@Setter
-public class ChatRoomUser {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "chat_room_user_id")
-    private Long id;
-
-    @Column(name = "is_creator")
-    private int isCreator;
-
-    @ManyToOne
-    @JoinColumn(name = "chat_room_id")
-    private ChatRoom chatRoom;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "workspace_id")
-    private Workspace workspace;
-
-    private Boolean newState = false;
-}
