--- src/main/java/com/teamkrews/chat/model/request/ChatRoomCreationRequest.java
@@ -1,13 +1,0 @@
-package com.teamkrews.chat.model.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class ChatRoomCreationRequest {
-    private String workspaceUUID;
-    private List<Long> userIds;
-}
