--- src/main/java/com/teamkrews/chat/model/request/MessageDTO.java
@@ -1,11 +1,0 @@
-package com.teamkrews.chat.model.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MessageDTO {
-    private Long senderId;
-    private String content;
-}
