--- src/main/java/com/teamkrews/chat/model/response/ChatRoomDetailResponse.java
@@ -1,12 +1,0 @@
-package com.teamkrews.chat.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRoomDetailResponse {
-    private Long chatRoomId;
-    private Long chatRoomUserId;
-    private UserInfoResponse targetUser;
-}
