--- src/main/java/com/teamkrews/chat/model/response/ChatRoomListResponse.java
@@ -1,12 +1,0 @@
-package com.teamkrews.chat.model.response;
-
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRoomListResponse {
-    private List<ChatRoomDetailResponse> chatRooms = new ArrayList<>();
-}
