--- src/main/java/com/teamkrews/chat/model/response/ChatRoomResponse.java
@@ -1,11 +1,0 @@
-package com.teamkrews.chat.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRoomResponse {
-    private Long chatRoomId;
-    private String workspaceUUID;
-}
