--- src/main/java/com/teamkrews/chat/model/response/MessageResponse.java
@@ -1,16 +1,0 @@
-package com.teamkrews.chat.model.response;
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class MessageResponse {
-
-    private Long senderId;
-
-    private String content;
-
-    private LocalDateTime createdAt;
-}
