--- src/main/java/com/teamkrews/chat/model/response/UserInfoResponse.java
@@ -1,11 +1,0 @@
-package com.teamkrews.chat.model.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserInfoResponse {
-    private String nickName;
-    private String profileImageUrl;
-}
