--- src/main/java/com/teamkrews/chat/repository/ChatRoomRepository.java
@@ -1,12 +1,0 @@
-package com.teamkrews.chat.repository;
-
-import com.teamkrews.chat.model.ChatRoom;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-    List<ChatRoom> findAll();
-}
