+++ src/main/java/com/teamkrews/chatroom/model/ChatRoomNewStateFalseDto.java
@@ -1,0 +1,15 @@
+package com.teamkrews.chatroom.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ChatRoomNewStateFalseDto {
+    public  Long chatRoomUserId;
+    public  final Boolean newState = Boolean.FALSE;
+
+    public ChatRoomNewStateFalseDto(Long chatRoomUserId) {
+        this.chatRoomUserId = chatRoomUserId;
+    }
+}
