+++ src/main/java/com/teamkrews/chatroom/model/request/ChatRoomCreationRequest.java
@@ -1,0 +1,13 @@
+package com.teamkrews.chatroom.model.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class ChatRoomCreationRequest {
+    private String workspaceUUID;
+    private List<Long> userIds;
+}
