+++ src/main/java/com/teamkrews/chatroom/model/response/ChatRoomDetailResponse.java
@@ -1,0 +1,12 @@
+package com.teamkrews.chatroom.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ChatRoomDetailResponse {
+    private Long chatRoomId;
+    private Long chatRoomUserId;
+    private UserInfoResponse targetUser;
+}
