+++ src/main/java/com/teamkrews/chatroom/model/response/ChatRoomListResponse.java
@@ -1,0 +1,12 @@
+package com.teamkrews.chatroom.model.response;
+
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ChatRoomListResponse {
+    private List<ChatRoomDetailResponse> chatRooms = new ArrayList<>();
+}
