+++ src/main/java/com/teamkrews/chatroom/model/response/UserInfoResponse.java
@@ -1,0 +1,11 @@
+package com.teamkrews.chatroom.model.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserInfoResponse {
+    private String nickName;
+    private String profileImageUrl;
+}
