+++ src/main/java/com/teamkrews/chatroom/repository/ChatRoomRepository.java
@@ -1,0 +1,12 @@
+package com.teamkrews.chatroom.repository;
+
+import com.teamkrews.chatroom.model.ChatRoom;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
+    List<ChatRoom> findAll();
+}
