+++ src/main/java/com/teamkrews/message/model/Message.java
@@ -1,0 +1,36 @@
+package com.teamkrews.message.model;
+
+import com.teamkrews.chatroom.model.ChatRoom;
+import jakarta.persistence.*;
+
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Entity
+@Table(name = "messages")
+@Getter
+@Setter
+@EntityListeners(AuditingEntityListener.class)
+public class Message {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "message_id")
+    private Long messageId;
+
+    @Column(name = "sender_id")
+    private Long senderId;
+
+    @Column(name = "content")
+    private String content;
+
+    @ManyToOne
+    @JoinColumn(name = "chat_room_id")
+    private ChatRoom chatRoom;
+
+    @Column(name = "created_at")
+    @CreatedDate
+    private LocalDateTime createdAt;
+}
