+++ src/main/java/com/teamkrews/message/model/request/MessageDTO.java
@@ -1,0 +1,11 @@
+package com.teamkrews.message.model.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MessageDTO {
+    private Long senderId;
+    private String content;
+}
