+++ src/main/java/com/teamkrews/message/model/response/MessageResponse.java
@@ -1,0 +1,16 @@
+package com.teamkrews.message.model.response;
+
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MessageResponse {
+
+    private Long senderId;
+
+    private String content;
+
+    private LocalDateTime createdAt;
+}
