+++ src/main/java/com/teamkrews/message/repository/MessageRepository.java
@@ -1,0 +1,14 @@
+package com.teamkrews.message.repository;
+
+import com.teamkrews.chatroom.model.ChatRoom;
+import com.teamkrews.message.model.Message;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface MessageRepository extends JpaRepository<Message, Long> {
+    List<Message> findAllByChatRoom(ChatRoom chatRoom);
+}
+
